Moved HostNavbar to components for consistant locations, and made requested updates.
